.70-create-snackbar-configuration-and-common-composable Minor updates when holding event selected to be deleted - Removed methods from viewmodel and implemented MutableState variable inside EventsScreen.kt - Comments updates on EventsScreen.kt - Updates ConfirmationAlertDialog.kt message on AlertDialog to accept a null value. - Added snackBarHostState on EventsScreen call on NavigationGraph.kt

+++ app/src/main/java/com/exercise/matipv2/ui/MainScreenViewModel.kt
@@ -13,7 +13,6 @@ import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.MutableStateFlow
-import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.launch
 import javax.inject.Inject
@@ -25,9 +24,6 @@ class MainScreenViewModel @Inject constructor (
 
     private val _uiState = MutableStateFlow(MainScreenState())
     val uiState = _uiState.asStateFlow()
-
-    private val _eventToDelete = MutableStateFlow<Event?>(null)
-    val eventToDelete: StateFlow<Event?> = _eventToDelete
 
     init {
         resetState()
@@ -130,14 +126,12 @@ class MainScreenViewModel @Inject constructor (
     * Delete Functions
     */
 
-    fun deleteEvent(event: Event) {
+    fun deleteEvent(event: Event?) {
         viewModelScope.launch(Dispatchers.IO) {
-            matipRepository.deleteEvent(event)
+            if (event != null) {
+                matipRepository.deleteEvent(event)
+            }
         }
-    }
-
-    fun setEventToDelete(event: Event?) {
-        _eventToDelete.value = event
     }
 
     /*
